first attempt at implementing the "F" formatting code
First primitive attempt at changing "ToString" so it supports the "F" formatting code.
The idea is that, by using this code, the user gets back the currency as string, with decimal
places according to the current currency, number formatting according to the thread culture
OR the specified IFormatProvider, followed by the currency name in English.