Skip serialization tests for .NET 8.0 or greater
This commit adds conditional compilation directives to skip certain serialization tests if the target framework is.NET 8.0 or greater. This is due to the fact that these tests use BinaryFormatter, which is not supported in .NET 8.0 or later as per Microsoft's guidance. The Serialize and Deserialize methods have been marked as obsolete.